third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-removal-expected.txt

CONSOLE WARNING: line 17: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: line 19: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/get.txt' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Tests the behavior of whitelisting origins and removing them later.

Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost 
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.


Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.


Testing: source origin: http://127.0.0.1:8000 destination origin: hTtP:LoCaLhOsT allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.


Testing: source origin: http://127.0.0.1:8000 destination origin: http: allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt'.
